+++ src/Resources/PaymentTypes/PaylaterInvoice.php
@@ -26,4 +26,5 @@ namespace UnzerSDK\Resources\PaymentTypes;
 
 class PaylaterInvoice extends BasePaymentType
 {
+    protected const SUPPORT_DIRECT_PAYMENT_CANCEL = true;
 }